+++ driven/storage/adapter.go
@@ -139,7 +139,7 @@ func (sa *Adapter) GetPolls(user *model.User, filter model.PollsFilter, filterBy
 	}
 
 	findOptions := options.Find()
-	findOptions.SetSort(bson.D{{"poll.status", 1}, {"_id", -1}})
+	findOptions.SetSort(bson.D{{Key: "poll.status", Value: 1}, {Key: "_id", Value: -1}})
 
 	if filter.Limit != nil {
 		findOptions.SetLimit(*filter.Limit)